post api update

+++ petpals-backend/src/main/java/com/app/petpals/service/PostService.java
@@ -1,12 +1,16 @@
 package com.app.petpals.service;
 
 import com.app.petpals.entity.Post;
+import com.app.petpals.entity.PostComment;
 import com.app.petpals.entity.User;
 import com.app.petpals.exception.PostDataException;
+import com.app.petpals.exception.PostLikeException;
 import com.app.petpals.exception.PostNotFoundException;
+import com.app.petpals.exception.UserUnauthorizedException;
 import com.app.petpals.payload.LikePostRequest;
 import com.app.petpals.payload.PostAddRequest;
 import com.app.petpals.payload.PostEditRequest;
+import com.app.petpals.repository.PostCommentRepository;
 import com.app.petpals.repository.PostRepository;
 import com.app.petpals.repository.UserRepository;
 import jakarta.transaction.Transactional;
@@ -23,6 +27,7 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserService userService;
     private final UserRepository userRepository;
+    private final PostCommentRepository postCommentRepository;
 
     public List<Post> getTest() {
         return postRepository.findAll();
@@ -48,9 +53,12 @@ public class PostService {
         return postRepository.save(post);
     }
 
-    public Post updatePost(String postId, PostEditRequest request) {
+    public Post updatePost(String postId, PostEditRequest request, String userId) {
         if (request.getTitle() == null) throw new PostDataException("Post title is required.");
         Post post = getPostById(postId);
+        if (!post.getCreator().getId().equals(userId)) {
+            throw new UserUnauthorizedException("Unauthorized action.");
+        }
         post.setTitle(request.getTitle());
         post.setDescription(request.getDescription());
         return postRepository.save(post);
@@ -66,12 +74,24 @@ public class PostService {
     public Post likePost(String postId, LikePostRequest request) {
         User user = userService.getById(request.getUserId());
         Post post = getPostById(postId);
-        if (!user.getLikedPosts().contains(post) && !post.getLikes().contains(user)) {
+        if (!user.getLikedPosts().contains(post) || !post.getLikes().contains(user)) {
             user.getLikedPosts().add(post);
             post.getLikes().add(user);
             userRepository.save(user);
             return postRepository.save(post);
-        } else throw new PostNotFoundException("temp");
+        } else throw new PostLikeException("This post is already liked.");
+    }
+
+    @Transactional
+    public Post removeLikePost(String postId, LikePostRequest request) {
+        User user = userService.getById(request.getUserId());
+        Post post = getPostById(postId);
+        if (user.getLikedPosts().contains(post) || post.getLikes().contains(user)) {
+            user.getLikedPosts().remove(post);
+            post.getLikes().remove(user);
+            userRepository.save(user);
+            return postRepository.save(post);
+        } else throw new PostLikeException("This post was not liked yet.");
     }
 
     public Post deletePostPicture(String postId) {
@@ -81,7 +101,24 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(String postId) {
+    public void deletePost(String postId, String userId) {
+        Post post = getPostById(postId);
+        if (!post.getCreator().getId().equals(userId)) {
+            throw new UserUnauthorizedException("Unauthorized action.");
+        }
+
+        for (User user : post.getLikes()) {
+            user.getLikedPosts().remove(post);
+        }
+        post.getLikes().clear();
+
+        for (PostComment comment : post.getComments()) {
+            for (User user : comment.getLikes()) {
+                user.getLikedComments().remove(comment);
+            }
+            comment.getLikes().clear();
+        }
+
         postRepository.deleteById(postId);
     }
 }
